# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel 11 homeyfy APIs Documents",
        "version": "1.0.0.1"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register API",
                "description": "This is User register API",
                "operationId": "Register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "user_name",
                                    "email",
                                    "phone",
                                    "password",
                                    "user_type"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "user_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "integer"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "user_type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User Registered Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "User Registered Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login API",
                "description": "This is the user login API",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/listing/listings": {
            "post": {
                "tags": [
                    "Listings"
                ],
                "summary": "Create a new listing",
                "description": "Create a new listing",
                "operationId": "3da444d93fc1b6a9515113d13cf06288",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "listing_title",
                                    "description",
                                    "listing_type",
                                    "address",
                                    "state",
                                    "city",
                                    "country"
                                ],
                                "properties": {
                                    "listing_title": {
                                        "type": "string",
                                        "example": "Beautiful Apartment"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A beautiful apartment in the city center."
                                    },
                                    "listing_type": {
                                        "type": "string",
                                        "example": "Apartment"
                                    },
                                    "listing_bedrooms": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "guests": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "beds": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "baths": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "listing_rooms": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "listing_size": {
                                        "type": "integer",
                                        "example": 1200
                                    },
                                    "listing_size_unit": {
                                        "type": "string",
                                        "example": "sqft"
                                    },
                                    "affiliate_booking_link": {
                                        "type": "string",
                                        "example": "https://example.com"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "suny pul RYK"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "Punjab"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "RYK"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Pakistan"
                                    },
                                    "area": {
                                        "type": "string",
                                        "example": "suny pul"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Listing saved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "There was an error saving the listing. Please try again."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Register",
            "description": "Register"
        },
        {
            "name": "Login",
            "description": "Login"
        },
        {
            "name": "Listings",
            "description": "Listings"
        }
    ]
}